# src/Luxus.Connect.Web/docs/openapi.yaml
openapi: 3.1.1
info:
  title: Luxus.Connect.Api | v1
  version: 1.0.0
servers:
  - url: http://localhost:8002
paths:
  /v1/pre-signed-urls/upload:
    post:
      tags:
        - PreSignedUrls
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePresignedUploadUrlInput'
          text/json:
            schema:
              $ref: '#/components/schemas/CreatePresignedUploadUrlInput'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreatePresignedUploadUrlInput'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PresignedUrlModel'
            application/json:
              schema:
                $ref: '#/components/schemas/PresignedUrlModel'
            text/json:
              schema:
                $ref: '#/components/schemas/PresignedUrlModel'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - Bearer: []
  /v1/pre-signed-urls/download:
    post:
      tags:
        - PreSignedUrls
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePresignedDownloadUrlInput'
          text/json:
            schema:
              $ref: '#/components/schemas/CreatePresignedDownloadUrlInput'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreatePresignedDownloadUrlInput'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PresignedUrlModel'
            application/json:
              schema:
                $ref: '#/components/schemas/PresignedUrlModel'
            text/json:
              schema:
                $ref: '#/components/schemas/PresignedUrlModel'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - Bearer: []
  /v1/stats/dashboard:
    get:
      tags:
        - Stats
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/DashboardStatsResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardStatsResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/DashboardStatsResponse'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - Bearer: []
  /v1/provider-invoices:
    get:
      tags:
        - ProviderInvoices
      parameters:
        - name: page_index
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type: integer
            format: int32
        - name: page_size
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type: integer
            format: int32
        - name: processing_month_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PagedResponseOfListProviderInvoiceResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponseOfListProviderInvoiceResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/PagedResponseOfListProviderInvoiceResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - Bearer: []
    post:
      tags:
        - ProviderInvoices
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderInvoiceImportRequestInput'
          text/json:
            schema:
              $ref: '#/components/schemas/ProviderInvoiceImportRequestInput'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ProviderInvoiceImportRequestInput'
        required: true
      responses:
        '201':
          description: Created
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/RequestProviderInvoiceImportResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/RequestProviderInvoiceImportResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/RequestProviderInvoiceImportResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - Bearer: []
  /v1/provider-invoices/{id}:
    get:
      tags:
        - ProviderInvoices
      operationId: ProviderInvoicesController.GetById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/GetProviderInvoiceResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/GetProviderInvoiceResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/GetProviderInvoiceResponse'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - Bearer: []
  /v1/providers:
    get:
      tags:
        - Providers
      parameters:
        - name: page_index
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type: integer
            format: int32
        - name: page_size
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PagedResponseOfListProvidersResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponseOfListProvidersResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/PagedResponseOfListProvidersResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - Bearer: []
    post:
      tags:
        - Providers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProviderInput'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateProviderInput'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateProviderInput'
        required: true
      responses:
        '201':
          description: Created
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/CreateProviderResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/CreateProviderResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/CreateProviderResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '409':
          description: Conflict
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - Bearer: []
...
  /v1/processing-months/{id}:
    get:
      tags:
        - ProcessingMonths
      operationId: ProcessingMonthsController.GetById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/GetProcessingMonthResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/GetProcessingMonthResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/GetProcessingMonthResponse'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - Bearer: []
...
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
...
    CreatePresignedDownloadUrlInput:
      required:
        - bucket_name
        - object_key
      type: object
      properties:
        bucket_name:
          type: string
        object_key:
          type: string
        expires_in_seconds:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - 'null'
            - integer
          format: int32
    CreatePresignedUploadUrlInput:
      required:
        - bucket_name
        - object_key
      type: object
      properties:
        bucket_name:
          type: string
        object_key:
          type: string
        content_type:
          type:
            - 'null'
            - string
        expires_in_seconds:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - 'null'
            - integer
          format: int32
...
    DashboardStatsResponse:
      required:
        - billing_cycles_count
        - customers_count
        - providers_count
        - provider_invoices_count
        - phone_lines_count
      type: object
      properties:
        billing_cycles_count:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: integer
          format: int32
        customers_count:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: integer
          format: int32
        providers_count:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: integer
          format: int32
        provider_invoices_count:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: integer
          format: int32
        phone_lines_count:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: integer
          format: int32
...
    PagedResponseOfListProviderInvoiceResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ListProviderInvoiceResponse'
        total_count:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: integer
          format: int64
...
    PresignedUrlModel:
      required:
        - url
        - http_method
        - expires_at_utc
      type: object
      properties:
        url:
          type: string
        http_method:
          type: string
        expires_at_utc:
          type: string
          format: date-time
...
  securitySchemes:
    Bearer:
      type: oauth2
      flows:
        password:
          tokenUrl: http://host.docker.internal:8081/realms/luxus/protocol/openid-connect/token
          scopes: {}
...